+++ oceanics-io-api/src/shared/middleware.ts
@@ -6,7 +6,48 @@ import { Logtail } from "@logtail/node";
 import { ILogtailLog } from "@logtail/types";
 
 import * as db from "./queries";
-import { Node, RequestContext, Query, ErrorDetail, HttpMethod } from "oceanics-io-api-wasm";
+import { Node, RequestContext, Query, ErrorDetail } from "oceanics-io-api-wasm";
+
+type ContextData = {
+    user?: Node
+    nodes: Node[]
+    label?: string
+    provider?: Node
+}
+
+// Type for handlers, before response processing
+export type ApiHandler = (event: {
+    data: ContextData
+    queryStringParameters: Record<string, string>
+    body?: string
+}) => Promise<{
+    statusCode: number;
+    // Stub type for generic entity transform object.
+    data?: Record<string, unknown> | {name: string, url: string}[];
+}>
+
+enum Authentication {
+    Bearer = "BearerAuth",
+    ApiKey = "ApiKeyAuth",
+    Basic = "BasicAuth"
+}
+
+enum HttpMethod {
+    POST = "POST",
+    PUT = "PUT",
+    OPTIONS = "OPTIONS",
+    QUERY = "QUERY",
+    DELETE = "DELETE",
+    GET ="GET",
+    HEAD = "HEAD"
+}
+
+// Predictable inbound headers
+type Headers = { 
+    authorization?: string;
+    ["x-api-key"]?: string;
+    [key: string]: string;
+}
 
 const logging = new Logtail(process.env.LOGTAIL_SOURCE_TOKEN??"");
 logging.use(async (log: ILogtailLog) => {
@@ -17,88 +58,13 @@ logging.use(async (log: ILogtailLog) => {
     }
 });
 
-// Types of Auth we handle
-enum Authentication {
-    Bearer = "BearerAuth",
-    ApiKey = "ApiKeyAuth",
-    Basic = "BasicAuth"
-}
-
-// Stub type for generic entity transform object.
-type Properties = { [key: string]: unknown };
-type Route = {name: string, url: string};
-
-// Type of request after going through middleware. 
-type ApiEvent = HandlerEvent & {
-    data: {
-      user?: Node;
-      nodes: Node[];
-      label?: string;
-      provider?: Node;
-    }
-}
- 
-// Response before being processed by middleware
-type ApiResponse = {
-    statusCode: number;
-    data?: Properties|Route[];
-}
-
-// Type for handlers, more succinct
-export type ApiHandler = (event: ApiEvent) => Promise<ApiResponse>
-
-// Handler lookup
-type HttpMethods = {
-    [key in HttpMethod]?: ApiHandler;
-}
-
-// Predictable inbound headers
-type Headers = { 
-    authorization?: string;
-    ["x-api-key"]?: string;
-    [key: string]: string;
-}
 
 // Securely store and compare passwords
 const hashPassword = (password: string, secret: string) =>
     crypto.pbkdf2Sync(password, secret, 100000, 64, "sha512").toString("hex");
 
-type QueryString = {left?: string, uuid?: string, right?: string};
-
-/**
- * Matching pattern based on basic auth information
- */
-const basicAuthClaim = ({ authorization="::" }: Headers) => {
-    const [email, password, secret] = authorization.split(":");
-    return { email, credential: hashPassword(password, secret) };
-}
-
-/**
- * Matching pattern based on bearer token authorization with JWT. Used in Auth, and to 
- * validate other APIs.
- */
-const bearerAuthClaim = ({ authorization="::" }: Headers) => {
-    const [, token] = authorization.split(":");
-    const { uuid } = jwt.verify(token, process.env.SIGNING_KEY??"") as JwtPayload;
-    return { uuid };
-}
-
-/**
- * ApiKey is used to match to a provider claim
- */
-const apiKeyClaim = ({ ["x-api-key"]: apiKey }: Headers) => {
-    return { apiKey }
-}
-
-// security protocols if any
-const authMethod = (pathSpec: unknown, httpMethod: HttpMethod) => {
-    const reduceMethods = (lookup: unknown, schema: unknown) => Object.assign(lookup, schema);
-    return pathSpec[httpMethod.toLowerCase()].security.reduce(reduceMethods, {}); 
-}
-
+// Make it a valid JSON response. Core API doesn't use other content types. 
 const transformResponse = ({extension="", data, statusCode, headers}) => {
-     
-    // Make it a valid JSON response. Core API doesn't use other content types. 
     return {
         statusCode,
         headers: {
@@ -107,6 +73,62 @@ const transformResponse = ({extension="", data, statusCode, headers}) => {
         },
         body: JSON.stringify(data)
     };
+}
+
+/**
+ * Matching pattern based on bearer token authorization with JWT. Used in Auth, and to 
+ * validate other APIs.
+ * 
+ * Have to assume anything with uuid is valid until query hits
+ */
+const bearerAuth = (context: RequestContext, { authorization="::" }: Headers) => {
+    context.auth = Authentication.Bearer;
+    const [, token] = authorization.split(":");
+    const { uuid } = jwt.verify(token, process.env.SIGNING_KEY??"") as JwtPayload;
+    if (!uuid)
+        throw Error("Bearer token claim missing .uuid");
+    return {
+        user: Node.user(JSON.stringify({ uuid }))
+    }
+}
+
+const basicAuth = async (context: RequestContext, { authorization="::" }: Headers) => {
+    context.auth = Authentication.Basic;
+    const [email, password, secret] = authorization.split(":");
+    const user = Node.user(JSON.stringify({ email, credential: hashPassword(password, secret) }));
+    return {user: await db.auth(user)}
+}
+
+const apiKeyAuth = (context: RequestContext, { ["x-api-key"]: apiKey }: Headers, body: string) => {
+    // Only for registration on /auth route
+    context.auth = Authentication.ApiKey;
+    const provider = Node.provider(JSON.stringify({apiKey}));
+    if (!provider.patternOnly().includes("apiKey"))
+        throw Error(`API key auth claim missing .apiKey`)
+    
+    // Works as existence check, because we strip blank strings
+    const {
+        email="",
+        password="",
+        secret=""
+    } = JSON.parse(body);
+    const user = Node.user(JSON.stringify({
+        email,
+        uuid: crypto.randomUUID(),
+        credential: hashPassword(password, secret)
+    }));
+    return {provider, user}
+}
+
+// security protocols if any
+const authMethod = (pathSpec: unknown, httpMethod: HttpMethod) => {
+    const reduceMethods = (lookup: unknown, schema: unknown) => Object.assign(lookup, schema);
+    const [security] = pathSpec[httpMethod.toLowerCase()].security.reduce(reduceMethods, {}); 
+    return {
+        [Authentication.Bearer]: bearerAuth,
+        [Authentication.ApiKey]: apiKeyAuth,
+        [Authentication.Basic]: basicAuth
+    }[security]; 
 }
 
 /**
@@ -124,7 +146,9 @@ const transformResponse = ({extension="", data, statusCode, headers}) => {
  * The side-effect is that it can be hard to tell the difference between a 404 and
  * 403 error. 
  */
-export function NetlifyRouter(methods: HttpMethods, pathSpec?: unknown): Handler {
+export function NetlifyRouter(methods: {
+    [key in HttpMethod]?: ApiHandler;
+}, pathSpec?: unknown): Handler {
    
     const _methods = {
         ...methods,
@@ -137,122 +161,42 @@ export function NetlifyRouter(methods: HttpMethods, pathSpec?: unknown): Handler
     /**
      * Return the actual bound handler.
      */
-    const NetlifyHandler = async function ({
+    const ApiHandler = async function ({
         httpMethod: _method,
         body,
         headers,
-        queryStringParameters,
-        ...request
+        queryStringParameters
     }: HandlerEvent) {
-        const context = new RequestContext(queryStringParameters as unknown as Query, _method);
-        const httpMethod: HttpMethod = new HttpMethod(_method);
-        const start = new Date();
+        const query = queryStringParameters as unknown as Query;
+        const httpMethod = _method as HttpMethod;
+        const context = new RequestContext(query, httpMethod);
         const handler: ApiHandler = _methods[httpMethod];
         if (typeof handler === "undefined") {
             const detail = ErrorDetail.invalidMethod();
-            logging.warn(`Invalid method`, transformLogLine({
-                httpMethod, 
-                statusCode: detail.statusCode, 
-                start
-            }));
+            logging.warn(`Invalid method`, context.log_line(detail.statusCode));
             return detail
         }
         
-        const security: string[] = authMethod(pathSpec, httpMethod); 
-
-        let user: Node;
-        let provider: Node;
-        if (Authentication.Bearer in security) {
-            let claim: { uuid: string, error?: string };
-            context.auth = Authentication.Bearer;
-            try {
-                claim = bearerAuthClaim(headers);
-                if (!claim.uuid) throw Error("Bearer token claim missing .uuid");
-            } catch ({message}) {
-                const detail = ErrorDetail.unauthorized();
-                logging.error(message, transformLogLine({
-                    httpMethod, 
-                    statusCode: detail.statusCode, 
-                    start,
-                    auth: context.auth as Authentication
-                }));
-                return detail
-            }
-            // Have to assume anything with uuid is valid until query hits
-            user = Node.user(JSON.stringify(claim));
-            
-        } else if (Authentication.Basic in security) {
-            context.auth = Authentication.Basic;
-            user = Node.user(JSON.stringify(basicAuthClaim(headers)));
-            try {
-                user = await db.auth(user);
-            } catch (error) {
-                const detail = ErrorDetail.unauthorized();
-                logging.error(error.message, transformLogLine({
-                    user,
-                    httpMethod, 
-                    statusCode: detail.statusCode, 
-                    start,
-                    auth: context.auth as Authentication
-                }));
-                return detail
-            }
-        } else if (Authentication.ApiKey in security) {
-            // Only for registration on /auth route
-            provider = Node.materialize(JSON.stringify(apiKeyClaim(headers)), "p", "Provider");
-            context.auth = Authentication.ApiKey;
-            // Works as existence check, because we strip blank strings
-            const {
-                email="",
-                password="",
-                secret=""
-            } = JSON.parse(body??"{}")
-            if (!provider.patternOnly().includes("apiKey")) {
-                const detail = ErrorDetail.unauthorized();
-                logging.error(`API key auth claim missing .apiKey`, transformLogLine({
-                    user: Node.user(JSON.stringify({ email })),
-                    httpMethod, 
-                    statusCode: detail.statusCode, 
-                    start,
-                    auth: context.auth as Authentication
-                }));
-                return detail
-            }
-            user = Node.user(JSON.stringify({
-                email,
-                uuid: crypto.randomUUID(),
-                credential: hashPassword(password, secret)
-            }));
-        } else {
-            // Shouldn't occur
+        let data: ContextData;
+        try {
+            data = authMethod(pathSpec, httpMethod)(context, headers, body??"{}");
+            data.nodes = Node.fromRequest(httpMethod, body, queryStringParameters);
+        } catch ({message}) {
+            const detail = ErrorDetail.unauthorized();
+            logging.error(message, context.log_line(detail.statusCode));
+            return detail
         }
-
-        // parse path into resources and make request to handler
-        const nodes = Node.fromRequest(httpMethod as Method, body, queryStringParameters as QueryString);
-
-        // Make it a valid JSON response. Core API doesn't use other content types.
         const response = await handler({
-            data: { 
-                user,
-                provider,
-                nodes
-            },
-            headers,
-            queryStringParameters,
-            body: undefined,
-            httpMethod: undefined,
-            ...request
+            data,
+            queryStringParameters
         }).then(transformResponse);
 
-        logging.info(`${httpMethod} response with ${response.statusCode}`, transformLogLine({
-            user,
-            httpMethod: httpMethod as Method, 
-            statusCode: response.statusCode, 
-            start,
-            auth: context.auth as Authentication
-        }));
+        logging.info(
+            `${httpMethod} response with ${response.statusCode}`, 
+            context.log_line(response.statusCode)
+        );
         return response;
     }
 
-    return NetlifyHandler;
+    return ApiHandler;
 }
